We process m_step rows sequentially, so all of m should not be used when computing tiling params.

With this change, llama1B prefill on M1 Mac in ExecuTorch improves performance by:

* ~50 tokens/sec --> ~188 tokens/sec [3.76x speedup]
* ~194 tokens/sec --> ~380 tokense/sec (with KleidiAI) [1.95x speedup]

Decode performance is the same (~90 tokens/sec).